refactor: include datasets in datasources
