--- docs/python_api/dataset/study_index_gwas_catalog.md
@@ -1,9 +1,0 @@
----
-title: Study Index (GWAS Catalog)
----
-
-::: otg.dataset.study_index_gwas_catalog.StudyIndexGWASCatalog
-
-## Schema
-
---8<-- "assets/schemas/study_index.md"
