--- docs/python_api/dataset/study_locus_gwas_catalog.md
@@ -1,9 +1,0 @@
----
-title: Study Locus (GWAS Catalog)
----
-
-::: otg.dataset.study_locus_gwas_catalog.StudyLocusGWASCatalog
-
-## Schema
-
---8<-- "assets/schemas/study_locus.md"
