+++ src/otg/datasource/gwas_catalog/study_index.py
@@ -7,18 +7,17 @@ from typing import TYPE_CHECKING
 import pyspark.sql.functions as f
 import pyspark.sql.types as t
 
+from otg.common.spark_helpers import column2camel_case
 from otg.common.utils import parse_efos
-from otg.dataset.study_index_gwas_catalog import (
-    StudyIndexGWASCatalog as StudyIndexGWASCatalogDataset,
-)
+from otg.dataset.study_index import StudyIndex
 
 if TYPE_CHECKING:
     from pyspark.sql import Column, DataFrame
 
 
 @dataclass
-class GWASCatalogStudyIndex:
-    """Study index from GWAS Catalog.
+class StudyIndexGWASCatalogParser:
+    """GWAS Catalog study index parser.
 
     The following information is harmonised from the GWAS Catalog:
 
@@ -121,7 +120,7 @@ class GWASCatalogStudyIndex:
                     exploded_ancestries.alias("ancestries"),
                     resolved_sample_count.alias("sample_count"),
                 ),
-                GWASCatalogStudyIndex._merge_ancestries_and_counts,
+                StudyIndexGWASCatalogParser._merge_ancestries_and_counts,
             ),
             f.array().cast(schema),
             lambda x, y: f.array_union(x, y),
@@ -131,7 +130,7 @@ class GWASCatalogStudyIndex:
         return f.aggregate(
             parsed_sample_size,
             unique_ancestries,
-            GWASCatalogStudyIndex._normalize_ancestries,
+            StudyIndexGWASCatalogParser._normalize_ancestries,
         )
 
     @staticmethod
@@ -230,17 +229,17 @@ class GWASCatalogStudyIndex:
 
     @classmethod
     def _parse_study_table(
-        cls: type[GWASCatalogStudyIndex], catalog_studies: DataFrame
-    ) -> StudyIndexGWASCatalogDataset:
+        cls: type[StudyIndexGWASCatalogParser], catalog_studies: DataFrame
+    ) -> StudyIndexGWASCatalog:
         """Harmonise GWASCatalog study table with `StudyIndex` schema.
 
         Args:
             catalog_studies (DataFrame): GWAS Catalog study table
 
         Returns:
-            StudyIndexGWASCatalogDataset: Parsed and annotated GWAS Catalog study table.
+            StudyIndexGWASCatalog: Parsed and annotated GWAS Catalog study table.
         """
-        return StudyIndexGWASCatalogDataset(
+        return StudyIndexGWASCatalog(
             _df=catalog_studies.select(
                 f.coalesce(
                     f.col("STUDY ACCESSION"), f.monotonically_increasing_id()
@@ -261,16 +260,16 @@ class GWASCatalogStudyIndex:
                     "backgroundTraitFromSourceMappedIds"
                 ),
             ),
-            _schema=StudyIndexGWASCatalogDataset.get_schema(),
+            _schema=StudyIndexGWASCatalog.get_schema(),
         )
 
     @classmethod
     def from_source(
-        cls: type[GWASCatalogStudyIndex],
+        cls: type[StudyIndexGWASCatalogParser],
         catalog_studies: DataFrame,
         ancestry_file: DataFrame,
         sumstats_lut: DataFrame,
-    ) -> StudyIndexGWASCatalogDataset:
+    ) -> StudyIndexGWASCatalog:
         """Ingests study level metadata from the GWAS Catalog.
 
         Args:
@@ -279,7 +278,7 @@ class GWASCatalogStudyIndex:
             sumstats_lut (DataFrame): GWAS Catalog summary statistics list.
 
         Returns:
-            StudyIndexGWASCatalogDataset: Parsed and annotated GWAS Catalog study table.
+            StudyIndexGWASCatalog: Parsed and annotated GWAS Catalog study table.
         """
         # Read GWAS Catalogue raw data
         return (
@@ -288,3 +287,266 @@ class GWASCatalogStudyIndex:
             .annotate_sumstats_info(sumstats_lut)
             .annotate_discovery_sample_sizes()
         )
+
+
+@dataclass
+class StudyIndexGWASCatalog(StudyIndex):
+    """Study index dataset from GWAS Catalog.
+
+    A study index dataset captures all the metadata for all studies including GWAS and Molecular QTL.
+    """
+
+    def update_study_id(
+        self: StudyIndexGWASCatalog, study_annotation: DataFrame
+    ) -> StudyIndexGWASCatalog:
+        """Update studyId with a dataframe containing study.
+
+        Args:
+            study_annotation (DataFrame): Dataframe containing `updatedStudyId`, `traitFromSource`, `traitFromSourceMappedIds` and key column `studyId`.
+
+        Returns:
+            StudyIndexGWASCatalog: Updated study table.
+        """
+        self.df = (
+            self._df.join(
+                study_annotation.select(
+                    *[
+                        f.col(c).alias(f"updated{c}")
+                        if c not in ["studyId", "updatedStudyId"]
+                        else f.col(c)
+                        for c in study_annotation.columns
+                    ]
+                ),
+                on="studyId",
+                how="left",
+            )
+            .withColumn(
+                "studyId",
+                f.coalesce(f.col("updatedStudyId"), f.col("studyId")),
+            )
+            .withColumn(
+                "traitFromSource",
+                f.coalesce(f.col("updatedtraitFromSource"), f.col("traitFromSource")),
+            )
+            .withColumn(
+                "traitFromSourceMappedIds",
+                f.coalesce(
+                    f.col("updatedtraitFromSourceMappedIds"),
+                    f.col("traitFromSourceMappedIds"),
+                ),
+            )
+            .select(self._df.columns)
+        )
+
+        return self
+
+    def annotate_ancestries(
+        self: StudyIndexGWASCatalog, ancestry_lut: DataFrame
+    ) -> StudyIndexGWASCatalog:
+        """Extracting sample sizes and ancestry information.
+
+        This function parses the ancestry data. Also get counts for the europeans in the same
+        discovery stage.
+
+        Args:
+            ancestry_lut (DataFrame): Ancestry table as downloaded from the GWAS Catalog
+
+        Returns:
+            StudyIndexGWASCatalog: Slimmed and cleaned version of the ancestry annotation.
+        """
+        from otg.datasource.gwas_catalog.study_index import (
+            StudyIndexGWASCatalogParser as GWASCatalogStudyIndexParser,
+        )
+
+        ancestry = (
+            ancestry_lut
+            # Convert column headers to camelcase:
+            .transform(
+                lambda df: df.select(
+                    *[f.expr(column2camel_case(x)) for x in df.columns]
+                )
+            ).withColumnRenamed(
+                "studyAccession", "studyId"
+            )  # studyId has not been split yet
+        )
+
+        # Parsing cohort information:
+        cohorts = ancestry_lut.select(
+            f.col("STUDY ACCESSION").alias("studyId"),
+            GWASCatalogStudyIndexParser.parse_cohorts(f.col("COHORT(S)")).alias(
+                "cohorts"
+            ),
+        ).distinct()
+
+        # Get a high resolution dataset on experimental stage:
+        ancestry_stages = (
+            ancestry.groupBy("studyId")
+            .pivot("stage")
+            .agg(
+                f.collect_set(
+                    f.struct(
+                        f.col("broadAncestralCategory").alias("ancestry"),
+                        f.col("numberOfIndividuals")
+                        .cast(t.LongType())
+                        .alias("sampleSize"),
+                    )
+                )
+            )
+            .withColumn(
+                "discoverySamples",
+                GWASCatalogStudyIndexParser._parse_discovery_samples(f.col("initial")),
+            )
+            .withColumnRenamed("replication", "replicationSamples")
+            # Mapping discovery stage ancestries to LD reference:
+            .withColumn(
+                "ldPopulationStructure",
+                self.aggregate_and_map_ancestries(f.col("discoverySamples")),
+            )
+            .drop("initial")
+            .persist()
+        )
+
+        # Generate information on the ancestry composition of the discovery stage, and calculate
+        # the proportion of the Europeans:
+        europeans_deconvoluted = (
+            ancestry
+            # Focus on discovery stage:
+            .filter(f.col("stage") == "initial")
+            # Sorting ancestries if European:
+            .withColumn(
+                "ancestryFlag",
+                # Excluding finnish:
+                f.when(
+                    f.col("initialSampleDescription").contains("Finnish"),
+                    f.lit("other"),
+                )
+                # Excluding Icelandic population:
+                .when(
+                    f.col("initialSampleDescription").contains("Icelandic"),
+                    f.lit("other"),
+                )
+                # Including European ancestry:
+                .when(f.col("broadAncestralCategory") == "European", f.lit("european"))
+                # Exclude all other population:
+                .otherwise("other"),
+            )
+            # Grouping by study accession and initial sample description:
+            .groupBy("studyId")
+            .pivot("ancestryFlag")
+            .agg(
+                # Summarizing sample sizes for all ancestries:
+                f.sum(f.col("numberOfIndividuals"))
+            )
+            # Do arithmetics to make sure we have the right proportion of european in the set:
+            .withColumn(
+                "initialSampleCountEuropean",
+                f.when(f.col("european").isNull(), f.lit(0)).otherwise(
+                    f.col("european")
+                ),
+            )
+            .withColumn(
+                "initialSampleCountOther",
+                f.when(f.col("other").isNull(), f.lit(0)).otherwise(f.col("other")),
+            )
+            .withColumn(
+                "initialSampleCount",
+                f.col("initialSampleCountEuropean") + f.col("other"),
+            )
+            .drop(
+                "european",
+                "other",
+                "initialSampleCount",
+                "initialSampleCountEuropean",
+                "initialSampleCountOther",
+            )
+        )
+
+        parsed_ancestry_lut = ancestry_stages.join(
+            europeans_deconvoluted, on="studyId", how="outer"
+        )
+
+        self.df = self.df.join(parsed_ancestry_lut, on="studyId", how="left").join(
+            cohorts, on="studyId", how="left"
+        )
+        return self
+
+    def annotate_sumstats_info(
+        self: StudyIndexGWASCatalog, sumstats_lut: DataFrame
+    ) -> StudyIndexGWASCatalog:
+        """Annotate summary stat locations.
+
+        Args:
+            sumstats_lut (DataFrame): listing GWAS Catalog summary stats paths
+
+        Returns:
+            StudyIndexGWASCatalog: including `summarystatsLocation` and `hasSumstats` columns
+        """
+        gwas_sumstats_base_uri = (
+            "ftp://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/"
+        )
+
+        parsed_sumstats_lut = sumstats_lut.withColumn(
+            "summarystatsLocation",
+            f.concat(
+                f.lit(gwas_sumstats_base_uri),
+                f.regexp_replace(f.col("_c0"), r"^\.\/", ""),
+            ),
+        ).select(
+            f.regexp_extract(f.col("summarystatsLocation"), r"\/(GCST\d+)\/", 1).alias(
+                "studyId"
+            ),
+            "summarystatsLocation",
+            f.lit(True).alias("hasSumstats"),
+        )
+
+        self.df = (
+            self.df.drop("hasSumstats")
+            .join(parsed_sumstats_lut, on="studyId", how="left")
+            .withColumn("hasSumstats", f.coalesce(f.col("hasSumstats"), f.lit(False)))
+        )
+        return self
+
+    def annotate_discovery_sample_sizes(
+        self: StudyIndexGWASCatalog,
+    ) -> StudyIndexGWASCatalog:
+        """Extract the sample size of the discovery stage of the study as annotated in the GWAS Catalog.
+
+        For some studies that measure quantitative traits, nCases and nControls can't be extracted. Therefore, we assume these are 0.
+
+        Returns:
+            StudyIndexGWASCatalog: object with columns `nCases`, `nControls`, and `nSamples` per `studyId` correctly extracted.
+        """
+        sample_size_lut = (
+            self.df.select(
+                "studyId",
+                f.explode_outer(f.split(f.col("initialSampleSize"), r",\s+")).alias(
+                    "samples"
+                ),
+            )
+            # Extracting the sample size from the string:
+            .withColumn(
+                "sampleSize",
+                f.regexp_extract(
+                    f.regexp_replace(f.col("samples"), ",", ""), r"[0-9,]+", 0
+                ).cast(t.IntegerType()),
+            )
+            .select(
+                "studyId",
+                "sampleSize",
+                f.when(f.col("samples").contains("cases"), f.col("sampleSize"))
+                .otherwise(f.lit(0))
+                .alias("nCases"),
+                f.when(f.col("samples").contains("controls"), f.col("sampleSize"))
+                .otherwise(f.lit(0))
+                .alias("nControls"),
+            )
+            # Aggregating sample sizes for all ancestries:
+            .groupBy("studyId")  # studyId has not been split yet
+            .agg(
+                f.sum("nCases").alias("nCases"),
+                f.sum("nControls").alias("nControls"),
+                f.sum("sampleSize").alias("nSamples"),
+            )
+        )
+        self.df = self.df.join(sample_size_lut, on="studyId", how="left")
+        return self
