+++ tests/conftest.py
@@ -18,14 +18,14 @@ from otg.dataset.l2g_gold_standard import L2GGoldStandard
 from otg.dataset.l2g_prediction import L2GPrediction
 from otg.dataset.ld_index import LDIndex
 from otg.dataset.study_index import StudyIndex
-from otg.dataset.study_index_gwas_catalog import StudyIndexGWASCatalog
 from otg.dataset.study_locus import StudyLocus
-from otg.dataset.study_locus_gwas_catalog import StudyLocusGWASCatalog
 from otg.dataset.study_locus_overlap import StudyLocusOverlap
 from otg.dataset.summary_statistics import SummaryStatistics
 from otg.dataset.v2g import V2G
 from otg.dataset.variant_annotation import VariantAnnotation
 from otg.dataset.variant_index import VariantIndex
+from otg.datasource.gwas_catalog.associations import StudyLocusGWASCatalog
+from otg.datasource.gwas_catalog.study_index import StudyIndexGWASCatalog
 from utils.spark import get_spark_testing_conf
